requirements.txt
torch>=2.0.0
transformers>=4.30.0
peft>=0.5.0
deepspeed>=0.10.0
PyYAML>=6.0
datasets>=2.12.0
accelerate>=0.20.0 